requirements.txt
Django==2.2.4
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0
